📦 refactor(theming): drop setup and migration sections
