--- apps/web/public/documentation/theming/quick-start.md
@@ -1,19 +1,0 @@
-# Quick start (CLI)
-
-`zard-cli init` writes `src/styles.css` for you and asks which base color to use.
-
-```bash tab="npm" copyButton
-npx zard-cli@latest init
-```
-
-```bash tab="pnpm" copyButton
-pnpm dlx zard-cli@latest init
-```
-
-```bash tab="yarn" copyButton
-yarn zard-cli@latest init
-```
-
-```bash tab="bun" copyButton
-bunx zard-cli@latest init
-```
